Move overlays to subfolder
