+++ src/ArenaGS.Views/Scenes/Overlays/IOverlay.cs
@@ -1,7 +1,7 @@
 ﻿using ArenaGS.Views.Views;
 using SkiaSharp;
 
-namespace ArenaGS.Views.Scenes
+namespace ArenaGS.Views.Scenes.Overlays
 {
 	internal interface IOverlay
 	{
